+++ CHANGELOG.md
@@ -21,6 +21,7 @@
 - New fields in MicroActorData for "IsHostile", "IsKillable"
 - New fields in MicroActorData for "Detect gaps", "Detect obstacles"
 - Added Goals and Level advancing
+- Added Screen-shake to camera controller
 
 ## 0.5.4
 ### Fixed